Break enums and bitmasks out of item_properties
