+++ src/items/__init__.py
@@ -1,3 +1,4 @@
 from src.items.item import Item
-from src.items.item_properties import *
+from src.items.enums import *
+from src.items.bitmasks import *
 from src.items.weapon import OneHandedAxe, TwoHandedAxe
